Refactor estructura de autenticación: simplificar el flujo de inicio de sesión y registro, y agregar widgets de Login y Registration.
