+++ lib/pages/dashboard_page.dart
@@ -1,7 +1,6 @@
 // Copyright (c) 2025 HCC. All rights reserved.
 // Use of this source code is governed by an MIT-style license that can be
 // found in the LICENSE file.
-
 import 'package:flutter/material.dart';
 import 'package:hcc_app/pages/profile_page.dart';
 import 'package:hcc_app/widgets/hcc_app_bar.dart';
@@ -19,6 +18,9 @@ class _DashboardPageState extends State<DashboardPage> {
   int _selectedIndex = 0;
 
   late final List<Widget> _pages;
+  late final user = Provider.of<UserProvider>(context).firebaseUser;
+  late final userModel = Provider.of<UserProvider>(context).userModel;
+  late final userName = userModel?.name;
 
   @override
   void initState() {
@@ -49,13 +51,10 @@ class _DashboardPageState extends State<DashboardPage> {
 
   @override
   Widget build(BuildContext context) {
-    final userProvider = Provider.of<UserProvider>(context);
-    final userModel = userProvider.userModel;
-
     return Scaffold(
       appBar: HccAppBar(
-        user: userProvider.firebaseUser,
-        userName: userModel?.name,
+        user: user,
+        userName: userName,
         isDashboard: true,
         formattedDate: _getFormattedDate(),
       ),
